chore: mailer macro
